+++ demo/midgard-node/src/workers/utils/confirm-block-commitments.ts
@@ -1,0 +1,11 @@
+import { WorkerOutput as BlockCommitmentWorkerOutput } from "./commit-block-header.js";
+
+export type WorkerInput = {
+  data: BlockCommitmentWorkerOutput;
+};
+
+export type BlockConfirmationWorkerOutput = {
+  txSize: number;
+  mempoolTxsCount: number;
+  sizeOfBlocksTxs: number;
+};
